Implementation of TPC-H, support for MonetDB

+++ src/com/oltpbenchmark/benchmarks/tpcc/pojo/Item.java
@@ -23,7 +23,7 @@ public class Item {
 
 	public int i_id; // PRIMARY KEY
 	public int i_im_id;
-	public float i_price;
+    public double i_price;
 	public String i_name;
 	public String i_data;
